+++ webgraph/src/graphs/btree_graph.rs
@@ -193,7 +193,7 @@ impl<L: Clone + 'static> LabeledSequentialGraph<L> for LabeledBTreeGraph<L> {}
 
 impl<L: Clone + 'static> RandomAccessLabeling for LabeledBTreeGraph<L> {
     type Labels<'succ>
-        = LabeledSuccessors<'succ, L>
+        = LabeledSucc<'succ, L>
     where
         L: 'succ;
     #[inline(always)]
@@ -208,7 +208,7 @@ impl<L: Clone + 'static> RandomAccessLabeling for LabeledBTreeGraph<L> {
 
     #[inline(always)]
     fn labels(&self, node: usize) -> <Self as RandomAccessLabeling>::Labels<'_> {
-        LabeledSuccessors(self.succ[node].iter())
+        LabeledSucc(self.succ[node].iter())
     }
 }
 
@@ -359,7 +359,7 @@ impl SequentialLabeling for BTreeGraph {
 impl SequentialGraph for BTreeGraph {}
 
 impl RandomAccessLabeling for BTreeGraph {
-    type Labels<'succ> = Successors<'succ>;
+    type Labels<'succ> = Succ<'succ>;
 
     #[inline(always)]
     fn num_arcs(&self) -> u64 {
@@ -373,7 +373,7 @@ impl RandomAccessLabeling for BTreeGraph {
 
     #[inline(always)]
     fn labels(&self, node: usize) -> <Self as RandomAccessLabeling>::Labels<'_> {
-        Successors(self.0.succ[node].keys().copied())
+        Succ(self.0.succ[node].keys().copied())
     }
 }
 
@@ -387,13 +387,11 @@ impl From<LabeledBTreeGraph<()>> for BTreeGraph {
 
 #[doc(hidden)]
 #[repr(transparent)]
-pub struct LabeledSuccessors<'a, L: Clone + 'static>(
-    std::collections::btree_map::Iter<'a, usize, L>,
-);
+pub struct LabeledSucc<'a, L: Clone + 'static>(std::collections::btree_map::Iter<'a, usize, L>);
 
-unsafe impl<L: Clone + 'static> SortedIterator for LabeledSuccessors<'_, L> {}
+unsafe impl<L: Clone + 'static> SortedIterator for LabeledSucc<'_, L> {}
 
-impl<L: Clone + 'static> Iterator for LabeledSuccessors<'_, L> {
+impl<L: Clone + 'static> Iterator for LabeledSucc<'_, L> {
     type Item = (usize, L);
     #[inline(always)]
     fn next(&mut self) -> Option<Self::Item> {
@@ -401,7 +399,7 @@ impl<L: Clone + 'static> Iterator for LabeledSuccessors<'_, L> {
     }
 }
 
-impl<L: Clone + 'static> ExactSizeIterator for LabeledSuccessors<'_, L> {
+impl<L: Clone + 'static> ExactSizeIterator for LabeledSucc<'_, L> {
     #[inline(always)]
     fn len(&self) -> usize {
         self.0.len()
@@ -410,13 +408,11 @@ impl<L: Clone + 'static> ExactSizeIterator for LabeledSuccessors<'_, L> {
 
 #[doc(hidden)]
 #[repr(transparent)]
-pub struct Successors<'succ>(
-    core::iter::Copied<std::collections::btree_map::Keys<'succ, usize, ()>>,
-);
+pub struct Succ<'succ>(core::iter::Copied<std::collections::btree_map::Keys<'succ, usize, ()>>);
 
-unsafe impl SortedIterator for Successors<'_> {}
+unsafe impl SortedIterator for Succ<'_> {}
 
-impl Iterator for Successors<'_> {
+impl Iterator for Succ<'_> {
     type Item = usize;
     #[inline(always)]
     fn next(&mut self) -> Option<Self::Item> {
@@ -424,7 +420,7 @@ impl Iterator for Successors<'_> {
     }
 }
 
-impl ExactSizeIterator for Successors<'_> {
+impl ExactSizeIterator for Succ<'_> {
     #[inline(always)]
     fn len(&self) -> usize {
         self.0.len()